HvD: Updates to the NWChem MD jobs.
At the moment it looks as if reducing the time step by a factor 4
from 0.002 to 0.0005 solves the problem (note that I also put
Lennard-Jones potentials on the HO and HWS hydrogens but it is now
not clear whether that is really necessary, this can be tested but
that takes time). Note that the right time step is system dependent
as a potential with sharper features requires short time steps
to sample it appropriately.