Making a cyclic structure out of 1AAR does not work.
Basically you have to connect the loose end of one Ubiquitin to
LYS-48 of the other. Then you need to re-optimize the structure.
However, using OpenBabel rotamer searches does not get you anywhere
close. Trying an energy minimization does not work because the bond
that needs to be fixed is massively overstretched generating huge
forces that pull the system into totally unphysical territory. Also
the system does not have enough symmetry to try and use that to
construct a sensible system from just the Ubiquitin chain that is
already linked. Beyond these simple attempts to fix this solving this
problem becomes a research problem in its own right that we do not have
the time for. So we will just use the available 1AAR structure and
kick the rest to the curb.